feat: update field startegies
