+++ src/interface/simulation-studio/suite-generation.interface.ts
@@ -13,7 +13,7 @@ export enum SimulationType {
   INTEGRATION_TESTING = 'INTEGRATION_TESTING',
 }
 
-export type ContextFieldStrategyCode = 'keep_sample' | 'static' | 'range' | 'generated' | 'null' | 'skip';
+export type FieldStrategyCode = 'keep_sample' | 'static' | 'range' | 'skip' | 'random';
 
 export interface SuiteGeneration {
   id: number;
@@ -49,210 +49,6 @@ export interface CreateSuiteGenerationDto {
   generation_metadata?: Record<string, unknown>;
 }
 
-// ── Context TXTP Config ──────────────────────────────────────────────────────
-
-export interface SuiteContextTxtpConfig {
-  id: number;
-  generation_id: number;
-  txtp: string;
-  txtp_version: string;
-  display_order: number;
-  message_count: number;
-  schema_snapshot: Record<string, unknown>;
-  sample_payload_snapshot?: Record<string, unknown>;
-  faker_seed?: number;
-  generator_profile: Record<string, unknown>;
-  related_txtp_config_id?: number | null;
-  related_transaction?: string | null;
-  created_at: Date;
-}
-
-export interface CreateContextTxtpConfigDto {
-  generation_id: number;
-  txtp: string;
-  txtp_version: string;
-  display_order: number;
-  message_count: number;
-  schema_snapshot: Record<string, unknown>;
-  sample_payload_snapshot?: Record<string, unknown>;
-  faker_seed?: number;
-  generator_profile?: Record<string, unknown>;
-  related_txtp_config_id?: number | null;
-  related_transaction?: string | null;
-}
-
-export interface UpdateContextTxtpConfigDto {
-  message_count?: number;
-  faker_seed?: number;
-  generator_profile?: Record<string, unknown>;
-  related_txtp_config_id?: number | null;
-}
-
-export interface AddContextTxtpConfigDto {
-  txtp: string;
-  txtp_version: string;
-  message_count?: number;
-  related_context_txtp_id?: number;
-}
-
-// ── Context Field Strategies ─────────────────────────────────────────────────
-
-export interface ContextFieldStrategy {
-  id: number;
-  context_txtp_config_id: number;
-  field_path: string;
-  strategy_code: ContextFieldStrategyCode;
-  static_value?: unknown;
-  range_min?: number;
-  range_max?: number;
-  faker_semantic_type?: string;
-  generator_options: Record<string, unknown>;
-  is_required_override?: boolean;
-  created_at: Date;
-}
-
-export interface UpsertFieldStrategyDto {
-  field_path: string;
-  strategy_code: ContextFieldStrategyCode;
-  static_value?: unknown;
-  range_min?: number;
-  range_max?: number;
-  faker_semantic_type?: string;
-  generator_options?: Record<string, unknown>;
-  is_required_override?: boolean;
-}
-
-export interface ContextTxtpConfigWithStrategies {
-  context_txtp_config_id: number;
-  txtp: string;
-  txtp_version: string;
-  message_count: number;
-  display_order: number;
-  schema_snapshot: Record<string, unknown>;
-  sample_payload_snapshot?: Record<string, unknown>;
-  related_txtp_config_id?: number | null;
-  related_transaction?: string | null;
-  field_strategies: ContextFieldStrategy[];
-}
-
-export interface BulkConfigItemDto {
-  context_txtp_config_id: number;
-  message_count?: number;
-  faker_seed?: number;
-  generator_profile?: Record<string, unknown>;
-  field_strategies?: UpsertFieldStrategyDto[];
-}
-
-// ── Trigger TXTP Config ──────────────────────────────────────────────────────
-
-export type TriggerOverrideType = 'static' | 'range' | 'generated' | 'remove' | 'null';
-export type TriggerExpectedBand = 'good' | 'neutral' | 'bad' | 'error';
-
-export interface SuiteTriggerTxtpConfig {
-  id: number;
-  generation_id: number;
-  txtp: string;
-  txtp_version: string;
-  display_order: number;
-  message_count: number;
-  link_to_context_pairs: boolean;
-  payload_template_json: Record<string, unknown>;
-  expected_independent_variable?: number;
-  expected_result_band?: TriggerExpectedBand;
-  notes?: string;
-  faker_seed?: number;
-  generator_profile: Record<string, unknown>;
-  related_txtp_config_id?: number | null;
-  related_transaction?: string | null;
-  created_at: Date;
-}
-
-export interface CreateTriggerTxtpConfigDto {
-  generation_id: number;
-  txtp: string;
-  txtp_version: string;
-  display_order: number;
-  message_count: number;
-  payload_template_json: Record<string, unknown>;
-  link_to_context_pairs?: boolean;
-  expected_independent_variable?: number;
-  expected_result_band?: TriggerExpectedBand;
-  notes?: string;
-  faker_seed?: number;
-  generator_profile?: Record<string, unknown>;
-  related_txtp_config_id?: number | null;
-  related_transaction?: string | null;
-}
-
-export interface UpdateTriggerTxtpConfigDto {
-  message_count?: number;
-  link_to_context_pairs?: boolean;
-  payload_template_json?: Record<string, unknown>;
-  expected_independent_variable?: number;
-  expected_result_band?: TriggerExpectedBand;
-  notes?: string;
-  faker_seed?: number;
-  generator_profile?: Record<string, unknown>;
-  related_txtp_config_id?: number | null;
-}
-
-export interface AddTriggerTxtpConfigDto {
-  txtp: string;
-  txtp_version: string;
-  message_count?: number;
-  related_trigger_txtp_id?: number;
-}
-
-export interface TriggerFieldOverride {
-  id: number;
-  trigger_txtp_config_id: number;
-  field_path: string;
-  override_type: TriggerOverrideType;
-  static_value?: unknown;
-  range_min?: number;
-  range_max?: number;
-  faker_semantic_type?: string;
-  generator_options: Record<string, unknown>;
-  created_at: Date;
-}
-
-export interface UpsertTriggerFieldOverrideDto {
-  field_path: string;
-  override_type: TriggerOverrideType;
-  static_value?: unknown;
-  range_min?: number;
-  range_max?: number;
-  faker_semantic_type?: string;
-  generator_options?: Record<string, unknown>;
-}
-
-export interface TriggerTxtpConfigWithOverrides {
-  trigger_txtp_config_id: number;
-  txtp: string;
-  txtp_version: string;
-  message_count: number;
-  display_order: number;
-  payload_template_json: Record<string, unknown>;
-  link_to_context_pairs: boolean;
-  expected_result_band?: TriggerExpectedBand;
-  notes?: string;
-  related_txtp_config_id?: number | null;
-  field_overrides: TriggerFieldOverride[];
-  related_transaction?: string;
-}
-
-export interface BulkTriggerConfigItemDto {
-  trigger_txtp_config_id: number;
-  message_count?: number;
-  link_to_context_pairs?: boolean;
-  payload_template_json?: Record<string, unknown>;
-  expected_result_band?: TriggerExpectedBand;
-  notes?: string;
-  faker_seed?: number;
-  generator_profile?: Record<string, unknown>;
-  field_overrides?: UpsertTriggerFieldOverrideDto[];
-}
-
 // ── Context/Trigger Mapping ──────────────────────────────────────────────────
 
 export interface MappingPair {
@@ -278,10 +74,7 @@ export interface TxtpMappingParamsDto {
   relatedTxtpId: string;
 }
 
-// ── Enrichment Tables ────────────────────────────────────────────────────────
-
-export type EnrichmentFieldStrategyCode = 'static' | 'range' | 'generated' | 'null' | 'copy';
-
+// ── Enrichment Tables ───────────────────────────────────────────────────────
 export interface SuiteEnrichmentTable {
   id: number;
   generation_id: number;
